Input for leaving date
I've changed the input style to gather info for the leaving date. This will be used later when I display how many days are left until the trip.

For now I'm going to move on to getting the next api call to work to get the relevant future weather data.