Weatherbit api call
Here I've retrieved the data from the weatherbit api. The furthest in the future it's possible to get a forecast seems to only be 16 days.

Either way the api is retrieving data successfully and logging it to the console. I'll continue onto part 8 and come back to this later if necessary.