replaceSpaces function
Here I've written the function to replace the spaces from the user city input to '+' symbols (for use in the pixabay api call).

The api call hasn't been successful so far, due to an Access-Control-Allow-Origin error. I'll work on fixing this is the next commit. The replaceSpaces function is working though, after console logging the output it does what I need it to do.