Pixabay api request
The pixabay api request is now working. The query is also handling multi-word searches thanks to the replaceSpaces function (e.g New York City).

The url, for the picture of the city being searched, is being logged to the console. Next I'll work on storing all of the data from the various api requests in server.js, so I can then display it to the user.