Changed structure
I've commented out my original code for now, and have started working on restructuring it.

The problem I've had recently is that, although I'm retrieving the correct data from the 3 apis, they are not being retrieved in the correct order. This has lead to trying to update the ui with data that hasn't been retrieved yet.

To combat this problem, I've put each api call into it's on separate async function.

I've then written the handleSubmit function, also an async function, which then runs the 3 api call functions in order.

The data is now being retrieved in the correct order. I checked this by logging the numbers 1 2 3 along with the data from each api call (if 123 are printed in order, the api calls are running in the correct order).

The next challenge will be to store the data and update the UI after the data has been retrieved. I'll work on this in the next commit.