UI updated (slight issue)
I'm getting closer to making everything work as I want it to.

Everything is running as expected, but there is one issue that needs addressing.

After the first UI update, the Generate Trip button needs to be clicked twice before the UI updates properly, firstly duplicating the previous trip.

I'll investigate this in the next commit.